requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Markdown==3.4.1
psycopg2-binary==2.9.5
python-environ==0.4.54
pytz==2022.7
sqlparse==0.4.3
